+++ lib/active_admin/application.rb
@@ -60,6 +60,12 @@ module ActiveAdmin
     # Whether the batch actions are enabled or not
     inheritable_setting :batch_actions, true
 
+    # Whether filters are enabled
+    inheritable_setting :filters, true
+
+    # The namespace root.
+    inheritable_setting :root_to, 'dashboard#index'
+
     # Active Admin makes educated guesses when displaying objects, this is
     # the list of methods it tries calling in order
     setting :display_name_methods, [ :display_name,