+++ lib/active_admin/namespace.rb
@@ -164,26 +164,21 @@ module ActiveAdmin
       config.controller.active_admin_config = config
     end
 
-    def resource_dsl
-      @resource_dsl ||= ResourceDSL.new
-    end
-
     def parse_registration_block(config, &block)
-      resource_dsl.run_registration_block(config, &block)
-    end
-
-    def page_dsl
-      @page_dsl ||= PageDSL.new
+      ResourceDSL.new(config).run_registration_block(&block)
     end
 
     def parse_page_registration_block(config, &block)
-      page_dsl.run_registration_block(config, &block)
+      PageDSL.new(config).run_registration_block(&block)
     end
 
     # Creates a dashboard controller for this config
     def generate_dashboard_controller
-      eval "class ::#{dashboard_controller_name} < ActiveAdmin::Dashboards::DashboardController; end"
-    end
+      return unless ActiveAdmin::Dashboards.built?
 
+      eval "class ::#{dashboard_controller_name} < ActiveAdmin::PageController
+              include ActiveAdmin::Dashboards::DashboardController
+            end"
+    end
   end
 end