+++ spec/unit/views/components/paginated_collection_spec.rb
@@ -2,7 +2,6 @@ require 'spec_helper'
 
 describe ActiveAdmin::Views::PaginatedCollection do
   describe "creating with the dsl" do
-    setup_arbre_context!
 
     let(:collection) do
       posts = [Post.new(:title => "First Post"), Post.new(:title => "Second Post"), Post.new(:title => "Third Post")]
@@ -14,9 +13,18 @@ describe ActiveAdmin::Views::PaginatedCollection do
       reload_routes!
     end
 
-    before do
-      request.stub!(:query_parameters).and_return({:controller => 'admin/posts', :action => 'index', :page => '1'})
-      controller.params = {:controller => 'admin/posts', :action => 'index'}
+    let(:view) do
+      view = mock_action_view
+      view.request.stub!(:query_parameters).and_return({:controller => 'admin/posts', :action => 'index', :page => '1'})
+      view.controller.params = {:controller => 'admin/posts', :action => 'index'}
+      view
+    end
+
+    # Helper to render paginated collections within an arbre context
+    def paginated_collection(*args)
+      render_arbre_component({:paginated_collection_args => args}, view) do
+        paginated_collection(*paginated_collection_args)
+      end
     end
 
     context "when specifying collection" do
@@ -171,6 +179,20 @@ describe ActiveAdmin::Views::PaginatedCollection do
       it "should display proper message (including number and not hash)" do
         pagination.find_by_class('pagination_information').first.content.
           gsub('&nbsp;',' ').should == "Displaying posts <b>1 - 2</b> of <b>3</b> in total"
+      end
+    end
+
+    context "when viewing the last page of a collection that has multiple pages" do
+      let(:collection) do
+        posts = [Post.new] * 81
+        Kaminari.paginate_array(posts).page(3).per(30)
+      end
+
+      let(:pagination) { paginated_collection(collection) }
+
+      it "should show the proper item counts" do
+        pagination.find_by_class('pagination_information').first.content.
+            gsub('&nbsp;',' ').should == "Displaying posts <b>61 - 81</b> of <b>81</b> in total"
       end
     end